requirements/base.txt
asgiref>=3.8.1,<3.9
black>=24.8.0,<24.9
click>=8.1.7,<8.2
Django>=5.1.1,<5.2
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
drf-yasg>=1.21.7,<1.22
inflection>=0.5.1,<0.6
isort>=5.13.2,<5.14
mypy-extensions>=1.0.0,<1.2
packaging>=24.1,<24.2
pathspec>=0.12.1,<1.0
platformdirs>=4.3.6,<4.4
PyJWT>=2.9.0,<3.0
pytz>=2024.2,<2024.3
PyYAML>=6.0.2,<6.3
sqlparse>=0.5.1,<0.6
uritemplate>=4.1.1,<4.2
setuptools>=75.1.0,<76
django-phonenumber-field>=8.0.0,<9
psycopg2>=2.9.9,<3
psycopg2-binary>=2.9.9,<3
python-dotenv>=1.0.1
phonenumbers>=8.13.46,<8.14
django.contrib.postgres>=0.0.1,<1